chore(audio): remove manual DMA config for SAI1
In [stm32h7xx-hal issue 329][1] it was pointed out that the new "v1.1"
rev of the Daisy Seed switched ChannelA and ChannelB when it moved from
ak4556 codec to wm8731 codec.

The HAL was unable to support this mixing of channels and P2M and M2P
combination.

`0.14` added this feature in [PR 162][2] and so we can remove the need
to manually alter the direction of the DMA streams, and instead rely on
the types provided by the HAL.

Configuration was verified with [sai.cpp][3] and [daisey_seed.cpp][4] in
libDaisy.

Direction::RECEIVE  ChannelA DMA1_Stream0 DMA_PERIPH_TO_MEMORY
Direction::TRANSMIT ChannelB DMA1_Stream1 DMA_MEMORY_TO_PERIPH

**Note:** that this only works for Seeds Rev 1.1 boards since those all
have wm8731 codecs. Support for both codecs will be needed at some point.

[1]: https://github.com/stm32-rs/stm32h7xx-hal/issues/329
[2]: https://github.com/stm32-rs/stm32h7xx-hal/pull/162
[3]: https://github.com/electro-smith/libDaisy/blob/5f1d862c60c6966dda3fb8a71fceeb6aeec00e02/src/per/sai.cpp#L224-L279
[4]: https://github.com/electro-smith/libDaisy/blob/5f1d862c60c6966dda3fb8a71fceeb6aeec00e02/src/daisy_seed.cpp#L271-L307