fix(audio): setup DMA RAM in DRAM2 for all uses
This fixes scratchy audio.

The problem was not setting enough DRAM2 (in terms of length)
as DMA RAM (specific settings such as non-cacheable etc. are
in `init_dma`).

So rather than try and set each allocation up, let's just set a
block of 32kb for everything to use. "Everything" at this point
is only the input and output audio buffer, but in the future will
be the circular buffer used for ADC DMA reads.

If we need more we can increase that.